+++ CHANGELOG.md
@@ -65,6 +65,7 @@ pointer to avoid crashes in scripting due to invalid pointer ownership (#3781).
 - Exposed `TimeSeriesTable::trimToIndices` to public API. (#3824)
 - Fixed bug in `Logger::cout`, now it works at any logger level. (#3826)
 - Fixed bugs in `MocoCasOCProblem` and `CasOC::Problem` with incorrect string formatting. (#3828)
+- Fixed `MocoOrientationTrackingGoal::initializeOnModelImpl` to check for missing kinematic states, but allow other missing columns. (#3830)
 
 
 v4.5